services/order-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
python-dotenv
pydantic
requests
